api_domestica/requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
sqlalchemy
pydantic
passlib
python-jose
bcrypt